[LSP] Implement server-side Cromwell submissions
This commit enables submission of Run WDL command
through a Code Action exposed on the LSP server.

It moves all of the client-specific
logic for running the workflows to server-side,
in contrast to how it's been done for WDL Debugger extension.

This feature can then be used in an editor-agnostic way.
However, currently it relies on a proposed 'progress'
feature for [LSP]
(https://github.com/microsoft/vscode-languageserver-node/blob/master/protocol/src/protocol.progress.proposed.md),
which may not yet be implemented in most clients.

It should be possible to check if the client
supports this feature, and if not, either
1) implement progress functionality on the client; or
2) gracefully switch server-side notification mechanism
   to that of "regular" message updates.

Currently, only vscode-langugeclient@next seems to support it.