Implement in-memory parsing of WDL in LS
This commit uses the new MiniWDL API to
parse WDL from source_text, instead of the URI.

The text gets updated on every 'textDocument/didChange'
notification from the client, and then
validated asynchronously with a "debounce" of 1 sec.

This means that not every change will lead to re-validation,
but it will wait for 1 sec until no more changes are sent,
and then re-validate.

We also refactor some global objects into members of Server class.