Set POSIX PATH for ShellCheck in the language server
Currently, MiniWDL attempts to find ShellCheck
on system PATH. However, clients like IntelliJ
may not always pass the entire PATH to the server process.

For example, it doesn't pass /usr/local/bin on MacOS.
This means that even if ShellCheck is installed
using Homebrew, the server process is unable to find it.

This commit attempts to fix this situation
(at least, temporarily) by patching the PATH
environmental variable on POSIX systems with
/usr/local/bin as the most likely candidate,
directly inside the Python server process.

In the future, we should add wdl.shellCheckPath
configuration variable to resolve the path
deterministically.